Types structure updated
