+++ packages/singlestore-elegance-sdk/src/shared/types/error.ts
@@ -1,0 +1,4 @@
+export type DefaultError = {
+  message: string;
+  status?: number;
+};
