+++ packages/singlestore-elegance-sdk/src/shared/types/request.ts
@@ -1,13 +1,10 @@
-import type { ConnectionTypes, CreateFileEmbeddingsBody, CreateAndInsertFileEmbeddingsBody } from "../shared/types";
-import { createEleganceClient } from "./index";
+import { CreateAndInsertFileEmbeddingsBody, CreateFileEmbeddingsBody } from "./controllers";
+
+export type Request = (...args: any[]) => Promise<any>;
 
 export type RequestInit = Omit<Exclude<Parameters<typeof fetch>[1], undefined>, "body" | "method"> & {
   [K: string]: any;
 };
-
-export type ClientConfig = { baseURL: string; defaultRequestOptions?: Partial<RequestInit> };
-
-export type EleganceClient<T extends ConnectionTypes> = ReturnType<typeof createEleganceClient<T>>;
 
 export type CreateAndInsertFileEmbeddingsRequestBody = Omit<CreateAndInsertFileEmbeddingsBody, "dataURL"> &
   ({ dataURL: string; file?: never } | { file: File; dataURL?: never });
